app/requirements.txt
# Para STT con Whisper
accelerate>=0.12.0
scipy>=1.7.0
torch>=2.0.1
transformers>=4.30.0
fastapi==0.115.6
uvicorn[standard]==0.30.6
qdrant-client==1.16.2
sentence-transformers==3.0.1
langchain-huggingface==1.2.0
langchain-community
pypdf==5.0.1
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0
tqdm==4.66.5
numpy==2.3.5 #numpy>=2.4.1
tf_keras
langchain-openai
langchain-qdrant
python-telegram-bot
google-cloud-dialogflow
openai-whisper==20250625
markitdown